Simplify earned browse, intent bonus, extra time rate limiting, fix float duration
- Intent bonus: +10 min for setting intent during Free Time (EarnedBrowseManager)
- Extra time: backend-backed rate limiting (verified requests only, 2/day)
- Extra time: safe JS callbacks via JSONSerialization (MainWindow)
- Fix float durationMinutes: PlatformSession.durationMinutes now Double
  (extension sends float, Swift as? Int was returning nil → endsAt = nil)
- NativeMessagingHost: parse durationMinutes as Double with Int fallback
- Dashboard: show remaining extra time requests count
- Parallel dev: add worktree workflow docs to CLAUDE.md

+++ Intentional/TimeTracker.swift
@@ -80,7 +80,7 @@ class TimeTracker {
         var categories: [String]?
         var startedAt: Double?       // Date.now() timestamp (ms since epoch)
         var endsAt: Double?          // Date.now() + duration, or nil for unlimited
-        var durationMinutes: Int
+        var durationMinutes: Double
         var freeBrowse: Bool
 
         static func inactive() -> PlatformSession {
@@ -265,7 +265,7 @@ class TimeTracker {
                     startedAt: session.startTime,
                     endedAt: Date(),
                     durationSeconds: duration,
-                    plannedDurationSeconds: platformSession.durationMinutes > 0 ? platformSession.durationMinutes * 60 : nil,
+                    plannedDurationSeconds: platformSession.durationMinutes > 0 ? Int(platformSession.durationMinutes * 60) : nil,
                     videoSeconds: session.videoSeconds
                 )
             }
@@ -387,7 +387,7 @@ class TimeTracker {
     }
 
     /// Update an existing platform session (e.g. extend timer)
-    func updatePlatformSession(for platform: String, endsAt: Double?, durationMinutes: Int?) {
+    func updatePlatformSession(for platform: String, endsAt: Double?, durationMinutes: Double?) {
         let key = platform.lowercased()
         guard platformSessions[key]?.active == true else { return }
